nir/algebraic: Don't generate 64-bit iadd3
The only other driver that sets has_iadd3 is NAK, and that does not
appear to support 64-bit integers. As a result, this should not have any
impact there.

No shader-db or fossil-db changes.

v2: Add nor_lower_iadd3_64 flag so we can continue to generate 64-bit
iadd3 on NVIDIA platforms.