anv: Add support for compressed images allocation in Xe2
Xe2 replaces auxiliary surface mapping by software to compress buffers,
instead it reserves part of the memory for the compression purpose.

To enable compression in Xe2 it is necessary bind memory with one of
the PAT indexes that has compression enabled.
So image meet requirements in anv_image_supports_pat_compression()
it will ask application to allocate the memory for given image with
a special memory type that only supports compressed memory types.

It is still always returning false in anv_image_supports_pat_compression()
as it still needs more work before compression can be enabled but the
foundation for the compressed allocation is here.

Signed-off-by: José Roberto de Souza <jose.souza@intel.com>