panfrost: explicitly loop over surfaces
The surface iterator here doesn't really help much. It seems much more
to the point to explicitly iterate over the surfaces in the order the HW
descriptors expects it.

This will make it easier to change how things works in the future,
especially for V9 and later. More on that in a later commit.