panfrost: untangle faces from layers
It's really only on pre-V7 HW that we need special handling of cubemap
faces because of the strange order in the surface descriptors. So let's
move that hack into the pre-V7 section of the code.

This allows us to pass a few more piglits on V7 and later.

It should really be possible to whip the pre-V7 code-paths into
submission here as well, but so far I've been unsuccessful.