v3d: Use DRM_IOCTL_V3D_GET_COUNTER to get perfcnt information
Currently, the information about the performance counters is duplicated
both in the kernel and in user space. Naturally, this leads to
inconsistency, as the user space might be updated and while the kernel
isn't.

Aiming to turn the kernel as the "single source of truth", use
DRM_IOCTL_V3D_GET_COUNTER, when available, to get performance counter
information.

Signed-off-by: Maíra Canal <mcanal@igalia.com>